Interface to query Consulta_CPMRC

Returns the coordinate of a state in a SRS given the Province, Municipality and Cadastral Reference of the state.

Usage

get_coor(RC, SRS= 'Google', Province = '',Municipality = '')

Arguments

RC

Cadastral reference of the state.

SRS

The Spatial Reference System desired for the coordinates, if it is missed, the returned coordinates will be given in the Spatial Reference System used by Googlemaps ('EPSG:4326'). The set of the available SRS to use are available in the vector coordinates from this package. Apart from that, the function accept as a SRS argument the values "Google" and "Oficial". The first uses the SRS value used by Google maps ('EPSG:4326') while the latter uses the official SRS ('EPSG:4258') in Europe. If no SRS is given, the function by default choose the "Google" SRS.

Province

A string with the Province of the state. It is not case sensitive.

Municipality

A string with the Municipality of the state. It is not case sensitive.

Value

A list consisting on two numeric values: the coordinates (latitude, longitude) and the reference system in which they are expressed.

References

http://ovc.catastro.meh.es/ovcservweb/OVCSWLocalizacionRC/OVCCoordenadas.asmx?op=Consulta_CPMRC

See also

Author

Angel Delgado Panadero.

Examples

# using all the arguments
coordinates <- get_coor("13077A01800039", "EPSG:4230", "CIUDAD REAL", "SANTA CRUZ DE MUDELA")
print(coordinates)
#> $coord
#> [1] 38.619657 -3.456242
#> 
#> $SRS
#> [1] "EPSG:4230"
#> 
#
# only the cadastral reference
coordinates <- get_coor("9872023VH5797S")
print(coordinates)
#> $coord
#> [1] 38.640143 -3.463284
#> 
#> $SRS
#> [1] "EPSG:4326"
#> 

Site built with pkgdown 2.0.2.

Template rostemplate by dieghernan, based on Bootstrapious.