Interface to query Consulta_RCCOOR_Distancia

Returns the Cadastral Reference of the state as well as the address (municipality, street and number) given its coordinates and the Spatial Reference System. In case that there is not any cadastral reference in the exact point, the function returns all the cadastral references in a square 50 meters long centered in the current point.

Usage

near_rc(lat, lon, SRS = "Google")

Arguments

lat

Latitude coordinate.

lon

Longitude coortinate.

SRS

The Spatial Reference System used for the coordinates. There is a vector with all the allowed SRSs in the variable from the package called coordinates.

The set of the available SRS to use are available in the vector `coordinates` from this package. A part from that, the function accept as a SRS argument the values "Google" and "Oficial". The first uses the SRS value used by Google maps ('EPSG:4326') while the latter uses the oficial SRS ('EPSG:4258') in Europe. If no SRS is given, the function by default choose the "Google" SRS.

Value

A data.frame with the the addresses of the states and the cadastral references in the nearby area of the location given for every SRS requested.

References

http://ovc.catastro.meh.es/ovcservweb/OVCSWLocalizacionRC/OVCCoordenadas.asmx?op=Consulta_RCCOOR_Distancia

Author

Angel Delgado Panadero.

Examples

direction <- near_rc(40.96002, -5.663408)
print(direction)    
#>                                         address             RC       SRS
#> 1 AY SANTO DOMINGO, DE 24 SALAMANCA (SALAMANCA) 6078401TL7367G EPSG:4326

Site built with pkgdown 2.0.2.

Template rostemplate by dieghernan, based on Bootstrapious.