ATOM INSPIRE: Reference database for ATOM addresses

Create a database containing the URLs provided in the INSPIRE ATOM service of the Spanish Cadastre for extracting addresses.

catr_atom_get_address_db_all() provides a top-level table including information of all the territorial offices (except Basque Country and Navarre) listing the municipalities included on each office.

catr_atom_get_address_db_to() provides a table for the specified territorial office including information for each of the municipalities of that office.

Usage

catr_atom_get_address_db_all(
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)

catr_atom_get_address_db_to(
  to,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)

Source

https://www.catastro.hacienda.gob.es/INSPIRE/Addresses/ES.SDGC.AD.atom.xml

Arguments

cache

A logical whether to do caching. Default is TRUE. See About caching section on catr_set_cache_dir().

update_cache

A logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source file.

cache_dir

A path to a cache directory. On NULL value (the default) the function would store the cached files on the tempdir.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

to

Territorial office. It can be any type of string, the function would perform a search using base::grep().

Value

A tibble with the information requested.

catr_atom_get_address_db_all() includes the following fields:

territorial_office: Territorial office, corresponding to each province of Spain except the Basque Country and Navarre.

url: ATOM URL for the corresponding territorial office.

munic: Name of the municipality.

date: Reference date of the data. Note that the information from this service is updated twice a year.

catr_atom_get_address_db_to() includes the following fields:

munic: Name of the municipality.

url: URL for downloading information of the corresponding municipality.

date: Reference date of the data. Note that the information from this service is updated twice a year.

See also

Examples

# \donttest{
catr_atom_get_address_db_all()
#> # A tibble: 7,611 × 4
#>    territorial_office             url                  munic date               
#>    <chr>                          <chr>                <chr> <dttm>             
#>  1 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  2 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  3 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  4 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  5 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  6 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  7 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  8 Territorial office 02 Albacete http://www.catastro… 0200… 2025-08-08 00:00:00
#>  9 Territorial office 02 Albacete http://www.catastro… 0201… 2025-08-08 00:00:00
#> 10 Territorial office 02 Albacete http://www.catastro… 0201… 2025-08-08 00:00:00
#> # ℹ 7,601 more rows
# }

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on templatesherpa.com.