Client tool for WFS INSPIRE services

Client tool for WFS INSPIRE services

Usage

inspire_wfs_get(
  scheme = "https",
  hostname = "ovc.catastro.meh.es",
  path = "INSPIRE/wfsCP.aspx",
  query = list(),
  verbose = FALSE
)

Arguments

scheme

Identifies the protocol to be used to access the resource on the Internet.

hostname

Identifies the host that holds the resource.

path

Identifies the specific resource in the host that the web client wants to access.

query

A named list with the name and value of the arguments to query.

verbose

logical. If TRUE displays informational messages.

Value

A character string with the path of the resulting file in the tempfile() folder.

Details

This function is used internally in all the WFS calls. We expose it to make it available to other users and/or developers for accessing other cadastral or INSPIRE resources. See Examples.

See also

Examples

# Accessing the Cadastre of Navarra
# Try also https://ropenspain.github.io/CatastRoNav/

file_local <- inspire_wfs_get(
  hostname = "inspire.navarra.es",
  path = "services/BU/wfs",
  query = list(
    service = "WFS",
    request = "getfeature",
    typenames = "BU:Building",
    bbox = "609800,4740100,611000,4741300",
    SRSNAME = "EPSG:25830"
  )
)
#> Error in httr2::req_perform(get_header): Failed to perform HTTP request.
#> Caused by error in `curl::curl_fetch_memory()`:
#> ! Timeout was reached [inspire.navarra.es]:
#> Failed to connect to inspire.navarra.es port 443 after 134125 ms: Couldn't connect to server

if (!is.null(file_local)) {
  pamp <- sf::read_sf(file_local)

  library(ggplot2)
  ggplot(pamp) +
    geom_sf()
}
#> Error: object 'file_local' not found

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on Universal Theme for Hugo.